remove FuelSetup component
